requirements.txt
attrs==20.2.0
aws-cdk.assets==1.70.0
aws-cdk.aws-cloudwatch==1.70.0
aws-cdk.aws-ec2==1.70.0
aws-cdk.aws-efs==1.70.0
aws-cdk.aws-events==1.70.0
aws-cdk.aws-iam==1.70.0
aws-cdk.aws-kms==1.70.0
aws-cdk.aws-logs==1.70.0
aws-cdk.aws-s3==1.70.0
aws-cdk.aws-s3-assets==1.70.0
aws-cdk.aws-ssm==1.70.0
aws-cdk.cloud-assembly-schema==1.70.0
aws-cdk.core==1.70.0
aws-cdk.cx-api==1.70.0
aws-cdk.region-info==1.70.0
cattrs==1.0.0
constructs==3.2.3
jsii==1.13.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
